requirements.txt
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
jupyter
pandas
scipy
sympy
matplotlib
tqdm
SimpleITK

yapf
tensorflow-addons
keras-tuner